+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.0-next.22](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.9.0-next.21...@esri/calcite-components-react@2.9.0-next.22) (2024-05-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [2.9.0-next.21](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.9.0-next.20...@esri/calcite-components-react@2.9.0-next.21) (2024-05-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [2.9.0-next.20](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.9.0-next.19...@esri/calcite-components-react@2.9.0-next.20) (2024-05-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [2.9.0-next.19](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.9.0-next.18...@esri/calcite-components-react@2.9.0-next.19) (2024-05-23)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [2.9.0-next.18](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.9.0-next.17...@esri/calcite-components-react@2.9.0-next.18) (2024-05-21)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [2.8.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@2.8.2...@esri/calcite-components-react@2.8.3) (2024-05-21)
 
 ### Miscellaneous Chores